+++ src/Commands/CommandFactoryInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Yumerov\MaxiBot\Commands;
+
+interface CommandFactoryInterface
+{
+
+    public function create(string $class): ?CommandInterface;
+}