+++ src/Repository/QuoteRepositoryInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Yumerov\MaxiBot\Repository;
+
+use Yumerov\MaxiBot\DTO\QuoteInterface;
+
+interface QuoteRepositoryInterface
+{
+
+    public function getRandomQuote(): ?QuoteInterface;
+}