+++ src/Traits/HasDiscordTrait.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Yumerov\MaxiBot\Traits;
+
+use Discord\Discord;
+
+trait HasDiscordTrait
+{
+
+    protected Discord $discord;
+
+    public function setDiscord(Discord $discord): static
+    {
+        $this->discord = $discord;
+        return $this;
+    }
+}