+++ src/functions.php
@@ -1,5 +1,9 @@
 <?php
 
+/**
+ * Set of functions for debugging and quick prototyping
+ */
+
 use Psr\Log\LoggerInterface;
 use Yumerov\MaxiBot\ContainerLoader;
 
@@ -16,5 +20,14 @@ function dd($arg): void
  */
 function logger(): LoggerInterface
 {
-    return ContainerLoader::getInstance()->get('Logger');
+    return get('Logger');
+}
+
+/**
+ * @throws \Yumerov\MaxiBot\Exceptions\Exception
+ * @throws Exception
+ */
+function get(string $id)
+{
+    return ContainerLoader::getInstance()->get($id);
 }