+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Symfony DI
 - PHPMetrics tool
 - Coffee reaction to good morning messages
+- Slash command "quote" to get a random Bitcoin related quote
 
 ### Changed
 - Refactored to use Symfony DI handle the class instance creating